requirements.txt
d2to1>=0.2.10,<0.3
pbr>=0.5,<0.6
SQLAlchemy>=0.7.8,<0.7.99
amqplib>=0.6.1
anyjson>=0.2.4
argparse
eventlet>=0.9.17
kombu>=1.0.4
lxml>=2.3
WebOb==1.2.3
greenlet>=0.3.1
sqlalchemy-migrate>=0.7.2
Babel>=0.9.6
iso8601>=0.1.4
setuptools_git>=0.4
python-keystoneclient>=0.2.0
oslo.config>=1.1.0
Flask==0.9
pecan>=0.2.0
wsme>=0.5b1
